Adapt interface to landscape vs portrait modes
Have the inventory bar stick to the left or the bottom depending on
which aspect ratio is present, optimizing for the most free space
available.

Also make some smaller code optimizations.

TODO: Make the portrait mode inventory only stick to the bottom with
`bottom: 0` if the height is less than the total height of the stage and
inventory combined. Otherwise, remove that property so it doesn't jump
to the bottom of the viewport. Also, account of the inventory size when
showing the stage, so that the fixed position doesn't obscure anything.