Add room support via <template>
Use the `<template>` tag to support rendering different scenes or
"rooms".

My strategy here is to be as simple as possible and use only core
browser features. A tool may be able to generate this at some point, but
for now this is a 'target' output that works as statically as practical.

It's technically a SPA, but if all the data were encoded in a single
HTML file.

I'm preferring this over a JSON payload, since that would eventually
have to be translated into HTML, and I want to get the final target
implementation first.