Add crosshair cursor when over stage
This pays a little homage to old school point and clicks that would use
a crosshair as the cursor.

It also changes expectations a little: we know that this is a bit of a
different context than a regular web page. It also does *not* indicate
what type of element is being hovered over; normally we see if something
is "clickable" by the cursor. In this case, the game can be more
challenging by not giving this away.