Add text display component
Add a component that animates text sent to a property. The animation
is a simple typewriter effect. This is achieved by adding `span` tags
around each letter in the text box display so their visibility can be
animated.

Add an accessible text version as well, so screen readers will correctly
read the copy.

The styling supports both light and dark system themes.