Accept arbitrary template for game objects
Add flexibility to game object contents by allowing implementers to pass
in an arbitrary template tag id for the component to use.

Get rid of the tiny game object CSS file, since the only style that will
eventually need to stay is the `position: absolute`.

Clean up the formatting of text display code.