Start refactor of utility bar to custom element
Move utility bar to it's own custom element. Rename is actions menu to
be less ambiguous.

Move scripts to the head and remove `async`; `defer` is the only behavior we
want here.

TODO: Move events to the custom element, and start handling
items/inventory.

NOTE this element can't have a shadow DOM because the main stage needs
to know about the CSS custom variables to lay out correctly. Might
refactor this at a later date if I think through a better alternative,
but using this for now.