Add room portal component
Allow sub classes to tack-on event behavior to the base `pointerup`
interaction event. This is a simple list of callbacks to load.

An enhancement would be to wrap the callbacks in promises to make their
execution deterministic.

Refactor the load room function so it's a member of a parent class.

Rigorously guard game object text display calls, by allowing the game
object text display interactions to bail out if there's no content,
instead of either throwing an error or displaying fallback text.

Format JSDoc and add it to package.