Background image support
Add classes and data attributes depending on which room is currently
loaded for custom effects.

Abandon background images via CSS for now, since they don't move with
the scroll, which ruins the effect of objects in space.

For now - and maybe forever - have a regular image be the background at
the top of the scope. Sometimes simple is best.