Put all scripts in a single file
This isn't a big enough project to split into modules. At least for now.
Keep it in a single file, and we can expose methods from there.

Keep all the component styles as string literals inside the component.
It's not perfect, but it's better than the absolute path of the linked
files.